more fixes, use new parent pom and maven
