+++ .jdkw
@@ -1,4 +1,4 @@
 JDKW_RELEASE=latest
 JDKW_DIST=zulu
-JDKW_BUILD=8.38.0.13
-JDKW_VERSION=8.0.212
+JDKW_BUILD=11.54.23-ca
+JDKW_VERSION=11.0.14
