+++ maven/maven-wrapper.properties
@@ -1,5 +1,5 @@
 #Maven download properties
-#Wed Feb 10 15:52:42 PST 2021
-checksumAlgorithm=SHA1
-verifyDownload=true
+#Thu Jan 20 00:03:35 PST 2022
 distributionUrl=https\://repo1.maven.org/maven2/org/apache/maven/apache-maven/3.6.3/apache-maven-3.6.3-bin.zip
+verifyDownload=true
+checksumAlgorithm=SHA1
