+++ pom.xml
@@ -18,7 +18,7 @@
   <parent>
     <groupId>com.arpnetworking.build</groupId>
     <artifactId>arpnetworking-parent-pom</artifactId>
-    <version>2.0.6</version>
+    <version>3.0.2</version>
     <relativePath />
   </parent>
